IoUring: Always try to enable IORING_SETUP_SUBMIT_ALL
Motivation

When porting io_uring to the main branch we introduced the use of IORING_SETUP_SUBMIT_ALL which makes a lot of sense. Unfortunally we never ended up actually enable it as we guarded it with an ifdef that was always false due the usage of an outdated io_uring.h file.

Motifications:

- Always try to enable IORING_SETUP_SUBMIT_ALL, at worse it will be just ignored if the kernel does not support it.
- Update the io_uring.h file

Result:

IORING_SETUP_SUBMIT_ALL is used when possible